Update Groups hasChild() method
Before, hasChild() accepted an element or an index. It returned true if the
element existed at any index, or it returned true if any element existed at
the index. However, it could not check for the existence of an element at
an index.

In the previous commit, I added the hasIndex() method to return true if any
element existed at an index, and I was going to update hasChild() to return
true if an element existing at any index or if an element existed at a given
index.

Now, I've combined both. The hasChild() method accepts an element, an index,
or both. If given an element and an index, it'll return true if the element
exists at the index. If given an element only, it'll return true if the
element exists at any index. If given an index only, it'll return true if
any element exist at the index.