Fix control word's __toString() method
Before, the Element\Control\Word's __toString() method would always include the
control word's parameter and a trailing-space delimiter. When the parameter is
the default parameter, it's optional. When the control word is the last child,
the space delimiter is uncessary.

Now, the parameter is not included if its value is the control word's default
value, and the trailing-space is not included if the control word is the group's
last child.