Fix two-state control word default parameter value
Before, I forced the default value of the two-state control words like B, I,
Strike, Super, Ul, and V to 1 for convenience. However, in the RTF spec,
any value but 0 is considered true. Thus, null is considered a true value.
I fixed the class logic to consider null true instead of forcing a default
value of 1.